note in the broadcast example about the use of keys
