+++ examples/consensus-node.rs
@@ -10,15 +10,16 @@ extern crate hbbft;
 extern crate log;
 extern crate pairing;
 extern crate protobuf;
-extern crate rand;
 
 mod network;
 
-use docopt::Docopt;
-use network::node::Node;
 use std::collections::HashSet;
 use std::net::SocketAddr;
 use std::vec::Vec;
+
+use docopt::Docopt;
+
+use network::node::Node;
 
 const VERSION: &str = "0.1.0";
 const USAGE: &str = "
@@ -62,6 +63,11 @@ pub fn main() {
     env_logger::init();
     let args: Args = parse_args();
     println!("{:?}", args);
-    let node = Node::new(args.bind_address, args.remote_addresses, args.value);
+
+    let node = Node::new(
+        args.bind_address,
+        args.remote_addresses,
+        args.value,
+    );
     node.run().expect("Node failed");
 }
